Add custom eslint rule for functional setState pattern.
This adds in a eslint rule for disallowing the access of this.props or
this.state from within a setState call. This is probably not going to
100% keep us in the clear I believe you could set a variable in an upper
scope equal to this.state.something and pass that into the setState
call. This seems to work well though for the most part. First time
writing an eslint thing so feedback would be appreciated.

**Testing Instructions**
run `npm i && npm run lint` to hopefully see that we do not have any of
these errors.

Trigger the lint errors by accessing this.props, or this.state within a
this.setState call, and run the lint or use an IDE linting tool.